[refactor] move validation schemas to validation folder
