+++ frontend/src/components/newScreening/NewScreening.tsx
@@ -1,7 +1,7 @@
 import ErrorBox from "@/components/ErrorBox";
 import NewScreeningHeader from "@/components/newScreening/NewScreeningHeader";
 import ProcessingStatusCard from "@/components/newScreening/ProcessingStatusCard";
-import type { JobDescriptionInput } from "@/components/newScreening/UploadJobDescriptionSchema";
+import type { JobDescriptionInput } from "@/validations/UploadJobDescriptionSchema";
 import ScreeningProgressSteps from "@/components/newScreening/ScreeningProgressSteps";
 import type { StepStatus } from "@/types/newScreeningTypes";
 import UploadJobDescriptionCard from "@/components/newScreening/UploadJobDescriptionCard";
